Module cfx_executor::executive

Transaction Execution Entry: Manages the execution of transactions. It is responsible for receiving transactions, performing checks according to the Conflux specification, and submitting them to the execution engine.

Re-exports

pub use executed::Executed;

pub use execution_outcome::ExecutionError;

pub use execution_outcome::ExecutionOutcome;

pub use execution_outcome::TxDropError;

pub use transact_options::ChargeCollateral;

pub use transact_options::TransactOptions;

pub use transact_options::TransactSettings;

Modules

executed

execution_outcome

transact_options

Structs

ExecutiveContext

Transaction executor.

Functions

contract_address

gas_required_for