Module cfx_executor::observer

Observability Interface: Defines a trait for extending functionality. Extensions can implement this trait to observe detailed aspects of the execution process.

Enums

AddressPocket

Traits

AsTracer

CallTracer

CheckpointTracer

DrainTrace

ExecutiveObserver

InternalTransferTracer

This trait is used by executive to build traces.

OpcodeTracer

StorageTracer

TracerTrait