cfx_executor::stack

Trait Resumable

pub trait Resumable: Send {
    // Required method
    fn resume(self: Box<Self>, result: FrameResult) -> Box<dyn Executable>;
}

Resumable is a trait representing objects for resuming the execution of a frame, which is suspended due to the invocation of another frame. The caller frame is resumed when the callee frame returns.

Required Methods

fn resume(self: Box<Self>, result: FrameResult) -> Box<dyn Executable>