Module cfx_executor::state

Ledger State: Acts as a caching and checkpoint layer built upon semantically meaningful database interfaces for the execution. Ledger State: Acts as a caching and checkpoint layer built upon semantically meaningful database interfaces for the execution.

Re-exports

pub use state_object::COMMISSION_PRIVILEGE_SPECIAL_KEY;

pub use state_object::COMMISSION_PRIVILEGE_SPECIAL_KEY;

Structs

COMMISSION_PRIVILEGE_SPECIAL_KEY

If we set this key, it means every account has commission privilege.

State

A caching and checkpoint layer built upon semantically meaningful database interfaces, providing interfaces and logics for managing accounts and global statistics to the execution engine.

StateCommitResult

Enums

CleanupMode

Mode of dealing with null accounts.

Functions

distribute_pos_interest

Distribute PoS interest to the PoS committee according to their reward points. Return the rewarded PoW accounts and their rewarded interest.

initialize_cip107

Initialize CIP-107 for the whole system.

initialize_cip137

Initialize CIP-137 for the whole system.

initialize_or_update_dao_voted_params

settle_collateral_for_all

Charge and refund all the storage collaterals. The suicided addresses are skimmed because their collateral have been checked out. This function should only be called in post-processing of a transaction.

update_pos_status