cfx_executor

Module internal_contract

Conflux Internal Contracts: Implements Conflux’s builtin contracts.

Modules

cross_space_events

pos_internal_entries

Structs

AllParamsVoteCount

IndexStatus

InternalContractExec

InternalContractMap

InternalRefContext

The internal contracts need to access the context parameter directly, e.g., foo(env, spec). But foo(context.env(), context.spec()) will incur lifetime issue. The InternalRefContext contains the parameters required by the internal contracts.

ParamVoteCount

Resume

Traits

InterfaceTrait

InternalContractTrait

Native implementation of an internal contract.

SolidityEventTrait

Native implementation of a solidity-interface function.

Functions

block_hash_slot

decode_register_info

epoch_hash_slot

get_settled_param_vote_count

If the vote counts are not initialized, all counts will be zero, and the parameters will be unchanged.

get_settled_pos_staking_for_votes

initialize_internal_contract_accounts

is_call_create_sig

is_withdraw_sig

make_staking_events

settle_current_votes

Move the next vote counts into settled and reset the counts. set_pos_staking is for compatibility with the Testnet.

storage_point_prop

suicide

The Actual Implementation of suicide. The contract which has non zero collateral_for_storage cannot suicide, otherwise it will: