cfx_executor::observer

Trait OpcodeTracer

pub trait OpcodeTracer {
    // Provided methods
    fn do_trace_opcode(&self, _enabled: &mut bool) { ... }
    fn initialize_interp(&mut self, gas_limit: U256) { ... }
    fn step(&mut self, interp: &dyn InterpreterInfo) { ... }
    fn step_end(&mut self, interp: &dyn InterpreterInfo) { ... }
    fn log(&mut self, address: &Address, topics: &Vec<H256>, data: &[u8]) { ... }
    fn selfdestruct(
        &mut self,
        space: Space,
        contract: &Address,
        target: &Address,
        value: U256,
    ) { ... }
}

Provided Methods

fn do_trace_opcode(&self, _enabled: &mut bool)

fn initialize_interp(&mut self, gas_limit: U256)

Called before the interpreter is initialized.

fn step(&mut self, interp: &dyn InterpreterInfo)

Called on each step of the interpreter.

Information about the current execution, including the memory, stack and more is available on interp (see [Interpreter]).

fn step_end(&mut self, interp: &dyn InterpreterInfo)

Called after step when the instruction has been executed.

fn log(&mut self, address: &Address, topics: &Vec<H256>, data: &[u8])

Called when a log is emitted.

fn selfdestruct( &mut self, space: Space, contract: &Address, target: &Address, value: U256, )

Called when a contract has been self-destructed with funds transferred to target.