cfx_executor::stack

Function exec_main_frame

pub fn exec_main_frame<'a>(
    main_frame: FreshFrame<'a>,
    resources: RuntimeRes<'a>,
) -> Result<FrameResult>

The function operates in a loop, starting with the execution of the main frame. Upon each frame’s invocation, the caller is pushed onto the call stack, and the callee is executed. After a frame completes execution, the function retrieves the result, pops a frame from the stack, and continues execution with the results from the callee. The loop continues until the call stack is empty, indicating that the main frame has finished executing.