cfx_executor::stack

Trait Executable

pub trait Executable: Send {
    // Required method
    fn execute(
        self: Box<Self>,
        context: Context<'_>,
    ) -> DbResult<ExecutableOutcome>;
}

Executable is a trait representing an object that can be executed within a frame.

There are generally two ways to create an Executable:

In a new frame, an Executable is created by the make_executable function, which uses the frame’s input parameters (ActionParams) and local information (FrameLocal) to determine the appropriate executable action.

After the completion of a frame’s execution, an Executable may be created by the its caller frame’s Resumer (implementing the Resumable trait) based on the execution results.

Required Methods

fn execute(self: Box<Self>, context: Context<'_>) -> DbResult<ExecutableOutcome>